Revert "NN-5364: get adjudication data from AdjudicationsAPI instead of PrisonAPI"
This reverts commit 513e20745db71b20625ffe8d4c5a8cf633e44375.
