+++ helm_deploy/values-prod.yaml
@@ -27,7 +27,6 @@ generic-service:
     SERVICES_PRISONS-API_BASE-URL: https://api.prison.service.justice.gov.uk
     SERVICES_CASE-NOTES_BASE-URL: https://offender-case-notes.service.justice.gov.uk
     SERVICES_AP-OASYS-CONTEXT-API_BASE-URL: https://approved-premises-and-oasys.hmpps.service.justice.gov.uk
-    SERVICES_MANAGE-ADJUDICATIONS-API_BASE-URL: https://manage-adjudications-api.hmpps.service.justice.gov.uk
     SERVICES_GOV-UK-BANK-HOLIDAYS-API_BASE-URL: https://www.gov.uk
     SPRING_FLYWAY_LOCATIONS: classpath:db/migration/all,classpath:db/migration/all-except-integration
     LOG-CLIENT-CREDENTIALS-JWT-INFO: false
