--- src/main/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/client/AdjudicationsApiClient.kt
@@ -1,28 +1,0 @@
-package uk.gov.justice.digital.hmpps.approvedpremisesapi.client
-
-import com.fasterxml.jackson.databind.ObjectMapper
-import org.springframework.beans.factory.annotation.Qualifier
-import org.springframework.beans.factory.annotation.Value
-import org.springframework.data.redis.core.RedisTemplate
-import org.springframework.stereotype.Component
-import org.springframework.web.reactive.function.client.WebClient
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.model.adjudications.AdjudicationsPage
-
-@Component
-class AdjudicationsApiClient(
-  @Qualifier("adjudicationsApiWebClient") webClient: WebClient,
-  objectMapper: ObjectMapper,
-  redisTemplate: RedisTemplate<String, String>,
-  @Value("\${preemptive-cache-key-prefix}") preemptiveCacheKeyPrefix: String,
-) : BaseHMPPSClient(webClient, objectMapper, redisTemplate, preemptiveCacheKeyPrefix) {
-
-  fun getAdjudicationsPage(nomsNumber: String, offset: Int?, pageSize: Int) = getRequest<AdjudicationsPage> {
-    withHeader("Page-Limit", pageSize.toString())
-
-    if (offset != null) {
-      withHeader("Page-Offset", offset.toString())
-    }
-
-    path = "/adjudications/$nomsNumber/adjudications"
-  }
-}
