+++ src/test/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/factory/AgencyFactory.kt
@@ -2,7 +2,7 @@ package uk.gov.justice.digital.hmpps.approvedpremisesapi.factory
 
 import io.github.bluegroundltd.kfactory.Factory
 import io.github.bluegroundltd.kfactory.Yielded
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.model.adjudications.Agency
+import uk.gov.justice.digital.hmpps.approvedpremisesapi.model.prisonsapi.Agency
 
 class AgencyFactory : Factory<Agency> {
   private var agencyId: Yielded<String> = { "AGNCY" }
