+++ src/test/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/unit/service/OffenderServiceTest.kt
@@ -8,7 +8,6 @@ import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
 import org.springframework.http.HttpMethod
 import org.springframework.http.HttpStatus
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.client.AdjudicationsApiClient
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.client.ApOASysContextApiClient
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.client.CaseNotesClient
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.client.ClientResult
@@ -49,7 +48,6 @@ class OffenderServiceTest {
   private val mockPrisonsApiClient = mockk<PrisonsApiClient>()
   private val mockCaseNotesClient = mockk<CaseNotesClient>()
   private val mockApOASysContextApiClient = mockk<ApOASysContextApiClient>()
-  private val mockAdjudicationsApiClient = mockk<AdjudicationsApiClient>()
   private val prisonCaseNotesConfigBindingModel = PrisonCaseNotesConfigBindingModel().apply {
     lookbackDays = 30
     prisonApiPageSize = 2
@@ -74,7 +72,6 @@ class OffenderServiceTest {
     mockPrisonsApiClient,
     mockCaseNotesClient,
     mockApOASysContextApiClient,
-    mockAdjudicationsApiClient,
     prisonCaseNotesConfigBindingModel,
     adjudicationsConfigBindingModel,
   )
@@ -546,7 +543,7 @@ class OffenderServiceTest {
     val nomsNumber = "NOMS456"
 
     every {
-      mockAdjudicationsApiClient.getAdjudicationsPage(
+      mockPrisonsApiClient.getAdjudicationsPage(
         nomsNumber = nomsNumber,
         pageSize = 2,
         offset = 0,
@@ -561,7 +558,7 @@ class OffenderServiceTest {
     val nomsNumber = "NOMS456"
 
     every {
-      mockAdjudicationsApiClient.getAdjudicationsPage(
+      mockPrisonsApiClient.getAdjudicationsPage(
         nomsNumber = nomsNumber,
         pageSize = 2,
         offset = 0,
@@ -604,7 +601,7 @@ class OffenderServiceTest {
       .produce()
 
     every {
-      mockAdjudicationsApiClient.getAdjudicationsPage(
+      mockPrisonsApiClient.getAdjudicationsPage(
         nomsNumber = nomsNumber,
         pageSize = 2,
         offset = 0,
@@ -615,7 +612,7 @@ class OffenderServiceTest {
     )
 
     every {
-      mockAdjudicationsApiClient.getAdjudicationsPage(
+      mockPrisonsApiClient.getAdjudicationsPage(
         nomsNumber = nomsNumber,
         pageSize = 2,
         offset = 2,
