+++ src/test/resources/application-test.yml
@@ -71,8 +71,6 @@ services:
     base-url: http://localhost:#WIREMOCK_PORT
   ap-oasys-context-api:
     base-url: http://localhost:#WIREMOCK_PORT
-  manage-adjudications-api:
-    base-url: http://localhost:#WIREMOCK_PORT
   gov-uk-bank-holidays-api:
     base-url: http://localhost:#WIREMOCK_PORT
